Move calpoker decoding code to new source file
