+++ src/cmd/main.rs
@@ -36,6 +36,7 @@ use chia_gaming::common::types::{
     Error, GameID, Hash, PrivateKey, Program, Sha256Input, Timeout,
 };
 use chia_gaming::games::poker_collection;
+use chia_gaming::games::calpoker::make_cards;
 use chia_gaming::outside::{GameStart, GameType, ToLocalUI};
 use chia_gaming::peer_container::{
     FullCoinSetAdapter, GameCradle, SynchronousGameCradle, SynchronousGameCradleConfig,
@@ -183,109 +184,6 @@ lazy_static! {
     };
 }
 
-fn mergein(outer: &[i32], inner: &[i32], offset: i32) -> Vec<i32> {
-    if inner.is_empty() {
-        outer.to_vec()
-    } else {
-        let first = inner[0] + offset;
-        let mut res = Vec::new();
-        if outer.is_empty() {
-            res.push(first);
-            res.append(&mut mergein(&[], &inner[1..], offset));
-        } else if outer[0] <= first {
-            res.push(outer[0]);
-            res.append(&mut mergein(&outer[1..], inner, offset + 1));
-        } else {
-            res.push(first);
-            res.append(&mut mergein(outer, &inner[1..], offset));
-        }
-        res
-    }
-}
-
-fn mergeover(outer: &[i32], inner: &[i32], offset: i32) -> Vec<i32> {
-    if inner.is_empty() {
-        vec![]
-    } else {
-        let first = inner[0] + offset;
-        let mut res = Vec::new();
-        if outer.is_empty() {
-            res.push(first);
-            res.append(&mut mergeover(&[], &inner[1..], offset));
-        } else if outer[0] <= first {
-            return mergeover(&outer[1..], inner, offset + 1);
-        } else {
-            res.push(first);
-            res.append(&mut mergeover(outer, &inner[1..], offset));
-        }
-        res
-    }
-}
-
-// Pick numchoose things out of numcards options with randomness extracted from vals
-// returns (cards newvals).
-fn choose(numcards: usize, numchoose: usize, randomness: BigInt) -> (Vec<i32>, BigInt) {
-    if numchoose == 1 {
-        let (newrandomness, card) = divmod(randomness, numcards.to_bigint().unwrap());
-        (vec![card.to_i32().unwrap()], newrandomness)
-    } else {
-        let half = numchoose >> 1;
-        let (cards1, newrandomness2) = choose(numcards, half, randomness);
-        let (cards2, newrandomness3) = choose(numcards - half, numchoose - half, newrandomness2);
-        (mergein(&cards1, &cards2, 0), newrandomness3)
-    }
-}
-
-fn cards_to_hand(cards: &[i32]) -> Vec<(i32, i32)> {
-    cards
-        .iter()
-        .map(|val| {
-            let rank = val / 4;
-            let suit = val % 4;
-            if rank == 12 {
-                (1, suit + 1)
-            } else {
-                (rank + 2, suit + 1)
-            }
-        })
-        .collect()
-}
-
-// Does the same thing as make_cards so bob can see the card display that alice can see once
-// the message of alice' preimage (alice_hash) goes through.
-fn make_cards(
-    alice_hash: &[u8],
-    bob_hash: &[u8],
-    amount: Amount,
-) -> (Vec<(i32, i32)>, Vec<(i32, i32)>) {
-    let amount_bigint = amount.to_u64().to_bigint().unwrap();
-    let (_, mut amount_bytes) = amount_bigint.to_bytes_be();
-    if &amount_bytes == &[0] {
-        amount_bytes = vec![];
-    }
-    if amount_bytes[0] & 0x80 != 0 {
-        amount_bytes.insert(0, 0);
-    }
-    eprintln!(
-        "make cards with alice_hash {alice_hash:?} bob_hash {bob_hash:?} amount {amount_bytes:?}"
-    );
-    let rand_input = Sha256Input::Array(vec![
-        Sha256Input::Bytes(&alice_hash[..16]),
-        Sha256Input::Bytes(&bob_hash[..16]),
-        Sha256Input::Bytes(&amount_bytes),
-    ])
-    .hash()
-    .bytes()
-    .to_vec();
-    let randomness = BigInt::from_signed_bytes_be(&rand_input);
-    let (handa, newrandomness) = choose(52, 8, randomness);
-    let (handb, _) = choose(52 - 8, 8, newrandomness);
-    (
-        cards_to_hand(&handa),
-        cards_to_hand(&mergeover(&handa, &handb, 0)),
-    )
-}
-
 impl GameRunner {
     fn new() -> Result<Self, Error> {
         let mut allocator = AllocEncoder::new();
@@ -421,17 +319,17 @@ impl GameRunner {
         Value::Object(r)
     }
 
-    fn convert_cards(&mut self, card_list: NodePtr) -> Vec<(i32, i32)> {
+    fn convert_cards(&mut self, card_list: NodePtr) -> Vec<(usize, usize)> {
         if let Some(cards_nodeptrs) = proper_list(self.allocator.allocator(), card_list, true) {
             return cards_nodeptrs
                 .iter()
                 .filter_map(|elt| {
                     proper_list(self.allocator.allocator(), *elt, true).map(|card| {
-                        let rank: i32 = atom_from_clvm(&mut self.allocator, card[0])
-                            .and_then(i32_from_atom)
+                        let rank: usize = atom_from_clvm(&mut self.allocator, card[0])
+                            .and_then(usize_from_atom)
                             .unwrap_or_default();
-                        let suit: i32 = atom_from_clvm(&mut self.allocator, card[1])
-                            .and_then(i32_from_atom)
+                        let suit: usize = atom_from_clvm(&mut self.allocator, card[1])
+                            .and_then(usize_from_atom)
                             .unwrap_or_default();
                         (rank, suit)
                     })
@@ -478,7 +376,7 @@ impl GameRunner {
     fn player_readable(&mut self, id: bool) -> Result<Value, String> {
         match &self.play_states[id as usize] {
             PlayState::BeforeAlicePicks => {
-                let cardlist: Vec<Vec<(i32, i32)>> = if let Some(cardlist) = proper_list(
+                let cardlist: Vec<Vec<(usize, usize)>> = if let Some(cardlist) = proper_list(
                     self.allocator.allocator(),
                     self.local_uis[id as usize]
                         .opponent_readable_move
