+++ src/games/mod.rs
@@ -1,3 +1,5 @@
+pub mod calpoker;
+
 use crate::common::standard_coin::read_hex_puzzle;
 use crate::common::types::{AllocEncoder, Program};
 use crate::outside::GameType;
